git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/fire_blaze_merge@7490 70805675-57b6-444b-bb00-18ecffcb4be8

+++ modules/spinupdata.cpp
@@ -66,7 +66,7 @@ void GenericSpinupData::detrend_data() {
 			annual_mean[y] += data[y][d];
 		}
 		annual_mean[y] /= data[y].size();
-		year_number[y] = y;
+		year_number[y] = (double)y;
 	}
 
 	double a, b;
